Structure refactoring. Code design improvements.
