+++ NetworkUtil/Entities/ResultOperation.cs
